+++ src/adapter/web-standard/index.ts
@@ -93,10 +93,8 @@ export const WebStandardAdapter: ElysiaAdapter = {
 			fnLiteral +=
 				`const u=r.url,` +
 				`s=u.indexOf('/',${standardHostname ? 11 : 7}),` +
-				`qi=u.indexOf('?',s+1)\n` +
-				`let p\n` +
-				`if(qi===-1)p=u.substring(s)\n` +
-				`else p=u.substring(s, qi)\n`
+				`qi=u.indexOf('?',s+1),` +
+				`p=u.substring(s,qi===-1?undefined:qi)\n`
 
 			if (hasTrace) fnLiteral += `const id=randomId()\n`