+++ test/extends/error.test.ts
@@ -2,7 +2,8 @@
 import { Elysia, t } from '../../src'
 
 import { describe, expect, it } from 'bun:test'
-import { req } from '../utils'
+import { post, req } from '../utils'
+import z from 'zod'
 
 class CustomError extends Error {
 	constructor() {
@@ -89,5 +90,39 @@ describe('Error', () => {
 
 		expect(response.status).toBe(422)
 		expect(response.headers.get('content-type')).toBe('application/json')
+	})
+
+	it('validation error should handle Standard Schema with error.detail', async () => {
+		const sendOtpEmailSchema = z.object({
+			channel: z.literal('email'),
+			otpTo: z.email({ error: 'Must be a valid email address' })
+		})
+
+		const sendOtpSmsSchema = z.object({
+			channel: z.literal('sms'),
+			otpTo: z.e164({
+				error: 'Must be a valid phone number with country code'
+			})
+		})
+
+		const sendOtpSchema = z.discriminatedUnion('channel', [
+			sendOtpEmailSchema,
+			sendOtpSmsSchema
+		])
+
+		const app = new Elysia()
+			.onError(({ code, error, status }) => {
+				switch (code) {
+					case 'VALIDATION':
+						return error.detail(error.message)
+				}
+			})
+			.post('/', ({ body, set }) => 'ok', {
+				body: sendOtpSchema
+			})
+
+		const response = await app.handle(post('/', {}))
+
+		expect(response.status).toBe(422)
 	})
 })